app/the-fool/requirements.txt
django==1.10.5
pyyaml